refactor: extract FiltersSort
